Move `filesystem` outside of `nums.core.backends`
